+++ CHANGES.md
@@ -58,8 +58,7 @@
 
 <!-- Changes that affect Black's preview style -->
 
-- Remove redundant parentheses around a generator expression passed as the sole argument
-  to a call (#5304)
+- Remove redundant parentheses around generator expressions (#5304)
 - Preserve two blank lines before a top-level class starting inside a `# fmt: off` block
   after an import (#5238)
 - Fix unnecessary parentheses around short RHS expressions in indexed assignments like